requirements.txt
-i
argon2-cffi==18.1.0
arrow==0.12.1
blessed==1.14.2
boto3==1.7.31
botocore==1.10.31
brotlipy==0.7.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
confusable-homoglyphs==3.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
django-allauth==0.36.0
django-anymail==3.0
django-appconf==1.0.2
django-avatar==4.1.0
django-braces==1.13.0
django-classy-tags==0.8.0
django-compressor==2.2
django-cookie-law==2.0.1
django-environ==0.4.4
django-fiction-outlines-api==0.1.2
django-fiction-outlines==0.2.2
django-filter==2.0.0.dev1
django-foundation-formtags==0.1.0
django-libsass==0.7
django-model-utils==3.1.2
django-picklefield==1.0.0
django-q-sentry==0.1.1
django-q==0.9.4
django-redis==4.9.0
django-rest-framework-rules==1.0.0
django-storages==1.6.6
django-taggit==0.22.2
django-treebeard==4.3
django==2.1a1
djangorestframework==3.8.2
docutils==0.14
gevent==1.3.2.post0
greenlet==0.4.13;
idna==2.6
itypes==1.1.0
jinja2==2.10
jmespath==0.9.3
libsass==0.14.5
markdown==2.6.11
markupsafe==1.0
oauthlib==2.1.0
pillow==5.1.0
psycopg2==2.7.4
pwned-passwords-django==1.1
pycparser==2.18
pytest-runner==4.2
python-dateutil==2.7.3
python-slugify==1.2.5
python3-openid==3.1.0
pytz==2018.4
raven==6.9.0
rcssmin==1.0.6
redis==2.10.6
requests-oauthlib==0.8.0
requests==2.18.4
rjsmin==1.0.12
rules==1.3
s3transfer==0.1.13
six==1.11.0
unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
waitress==1.1.0
wcwidth==0.1.7
wheel==0.31.1
whitenoise==4.0b4
-e git+https://github.com/ubernostrum/django-registration.git@f957d2824cf77820287da2993fca44d6445c592d#egg=django-registration
